@@ -4,19 +4,19 @@
   "private": true,
   "dependencies": {
     "classnames": "2.3.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "web-vitals": "2.1.2"
   },
   "scripts": {
-    "start": "craco start",
+    "start": "cross-env PORT=6969 TARGET_ENV=local craco start",
     "start:storybook": "start-storybook -p 3001",
-    "test": "craco test --watchAll=false",
-    "test:watch": "craco test --watch",
-    "test:ci": "craco test --watchAll=false --runInBand",
-    "build": "craco build",
+    "test": "cross-env TARGET_ENV=local craco test --watchAll=false",
+    "test:watch": "cross-env TARGET_ENV=local craco test --watch",
+    "test:ci": "cross-env TARGET_ENV=local craco test --watchAll=false --runInBand",
+    "build": "cross-env TARGET_ENV=local craco build",
     "build:storybook": "build-storybook -o build-storybook",
-    "lint": "eslint src .storybook *.js",
+    "lint": "eslint src .storybook *.js --fix",
     "prepare": "husky install"
   },
   "browserslist": {
@@ -32,23 +32,25 @@
     ]
   },
   "devDependencies": {
-    "@craco/craco": "6.4.3",
-    "@storybook/addon-actions": "6.5.0-alpha.5",
-    "@storybook/addon-essentials": "6.5.0-alpha.5",
-    "@storybook/addon-links": "6.5.0-alpha.5",
-    "@storybook/builder-webpack5": "6.5.0-alpha.5",
-    "@storybook/manager-webpack5": "6.5.0-alpha.5",
-    "@storybook/node-logger": "6.5.0-alpha.5",
-    "@storybook/preset-create-react-app": "4.0.0",
-    "@storybook/react": "6.5.0-alpha.5",
+    "@craco/craco": "7.0.0",
+    "@storybook/addon-actions": "6.5.16",
+    "@storybook/addon-essentials": "6.5.16",
+    "@storybook/addon-links": "6.5.16",
+    "@storybook/builder-webpack5": "6.5.16",
+    "@storybook/manager-webpack5": "6.5.16",
+    "@storybook/node-logger": "6.5.16",
+    "@storybook/preset-create-react-app": "4.1.2",
+    "@storybook/react": "6.5.16",
     "@testing-library/jest-dom": "5.16.1",
     "@testing-library/react": "12.1.2",
     "@testing-library/user-event": "13.5.0",
     "@types/jest": "27.0.3",
     "@types/node": "17.0.6",
-    "@types/react": "17.0.38",
-    "@types/react-dom": "17.0.11",
+    "@types/react": "18.0.27",
+    "@types/react-dom": "18.0.10",
     "autoprefixer": "10.4.0",
+    "circular-dependency-plugin": "5.2.2",
+    "cross-env": "7.0.3",
     "cssnano": "5.0.14",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-prettier": "4.0.0",
