@@ -1,19 +1,15 @@
-import { StrictMode } from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import './index.css'
+import App from './App'
+import reportWebVitals from './reportWebVitals'
 
-ReactDOM.render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-  document.getElementById('root')
-);
+import { createRoot } from 'react-dom/client'
+const container = document.getElementById('root')
+const root = createRoot(container!)
+root.render(<App />)
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-if (process.env.NODE_ENV === 'development') {
-  reportWebVitals(console.log);
+if (process.env.TARGET_ENV === 'local') {
+  reportWebVitals(console.log)
 }
